abstract common methods to a shared location
